feat/user AND feat/post, feat/comment modified

+++ src/main/java/com/tave8/ottu/entity/Genre.java
@@ -17,7 +17,7 @@ import javax.persistence.Table;
 public class Genre {
     @Id
     @Column(name = "genre_idx")
-    private Long genreIdx;
+    private int genreIdx;
 
     @Column(name = "genre_name")
     private String genreName;
